+++ src/gwt/src/org/rstudio/studio/client/workbench/ui/PaneManager.java
@@ -401,7 +401,12 @@ public class PaneManager
          el = el.getParentElement();
          for (LogicalWindow window : panes_)
          {
-            if (el.equals(window.getActiveWidget().getElement()))
+            Widget activeWidget = window.getActiveWidget();
+            if (activeWidget == null)
+               continue;
+            
+            Element activeEl = activeWidget.getElement();
+            if (el.equals(activeEl))
             {
                targetWindow = window;
                break;
@@ -431,12 +436,22 @@ public class PaneManager
       else
       {
          LogicalWindow activeWindow = getActiveLogicalWindow();
-         if (getConsoleLogicalWindow().equals(activeWindow))
+         LogicalWindow consoleWindow = getConsoleLogicalWindow();
+         
+         if (consoleWindow.equals(activeWindow))
          {
             consoleTabPanel_.selectNextTab();
          }
          else
          {
+            // Ensure that the console window is activated
+            if (consoleWindow.getState().equals(WindowState.MINIMIZE))
+            {
+               WindowStateChangeEvent event =
+                     new WindowStateChangeEvent(WindowState.NORMAL);
+               consoleWindow.onWindowStateChange(event);
+            }
+            
             consoleTabPanel_.selectTab(consoleTabPanel_.getSelectedIndex());
          }
       }